+++ server/src/main/java/com/aticatac/server/components/ai/SearchNode.java
@@ -1,7 +1,6 @@
 package com.aticatac.server.components.ai;
 
 import com.aticatac.common.components.transform.Position;
-
 import java.util.ArrayList;
 
 /**
@@ -11,37 +10,37 @@ import java.util.ArrayList;
  * @author Dylan
  */
 class SearchNode extends Position {
-    /**
-     * All the SearchNodes connected to this SearchNode
-     */
-    private ArrayList<SearchNode> connectedNodes;
+  /**
+   * All the SearchNodes connected to this SearchNode
+   */
+  private ArrayList<SearchNode> connectedNodes;
 
-    /**
-     * Creates a new SearchNode at (x, y).
-     *
-     * @param x X co-ordinate position of this SearchNode
-     * @param y Y co-ordinate position of this SearchNode
-     */
-    SearchNode(double x, double y) {
-        super(x, y);
-        connectedNodes = new ArrayList<SearchNode>();
-    }
+  /**
+   * Creates a new SearchNode at (x, y).
+   *
+   * @param x X co-ordinate position of this SearchNode
+   * @param y Y co-ordinate position of this SearchNode
+   */
+  SearchNode(int x, int y) {
+    super(x, y);
+    connectedNodes = new ArrayList<SearchNode>();
+  }
 
-    /**
-     * Gets a list of SearchNodes that this SearchNode is connected to.
-     *
-     * @return An ArrayList of the connected SearchNodes
-     */
-    public ArrayList<SearchNode> getConnectedNodes() {
-        return connectedNodes;
-    }
+  /**
+   * Gets a list of SearchNodes that this SearchNode is connected to.
+   *
+   * @return An ArrayList of the connected SearchNodes
+   */
+  public ArrayList<SearchNode> getConnectedNodes() {
+    return connectedNodes;
+  }
 
-    /**
-     * Adds a connection to another SearchNode.
-     *
-     * @param node The SearchNode to connect to
-     */
-    public void addConnection(SearchNode node) {
-        connectedNodes.add(node);
-    }
+  /**
+   * Adds a connection to another SearchNode.
+   *
+   * @param node The SearchNode to connect to
+   */
+  public void addConnection(SearchNode node) {
+    connectedNodes.add(node);
+  }
 }
